updated django CI workflow
we don't have tests written yet, so this is not giving us any information and shows up as a big red X, when what people add code it looks broken but is possibly up to our current standard. 

want pull requests to be more readable. 

to roll back change, just uncomment out lines 28-30 in django.yaml